requirements.txt
aiogram>=3.4,<4.0
gspread>=6.1,<7.0
google-auth>=2.0,<3.0
python-dotenv>=1.0,<2.0
